+++ ChangeLog.rst
@@ -1,3 +1,10 @@
+Unreleased
+==========
+----
+
+* Fixes #409: IndexError on empty strong mark.
+
+
 2024.2.25
 =========
 ----